Remove the Python code-generation implementation
Closes #238.
