--- .github/workflows/pytest.yml
@@ -1,45 +1,0 @@
-name: Python Tests
-
-on:
-  push:
-    branches: [main]
-    paths:
-      - "scripts/*.py"
-      - "scripts/*.j2"
-      - "pyproject.toml"
-      - "uv.lock"
-      - ".github/workflows/pytest.yml"
-  pull_request:
-    branches: [main]
-    paths:
-      - "scripts/*.py"
-      - "scripts/*.j2"
-      - "pyproject.toml"
-      - "uv.lock"
-      - ".github/workflows/pytest.yml"
-
-env:
-  UV_FROZEN: 1
-
-jobs:
-  pytest:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v5
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v10.0.1
-        with:
-          enable-cache: true
-
-      - name: Install dependencies
-        run: uv sync
-
-      - name: Run Ruff check
-        run: uv run ruff check scripts/*.py
-
-      - name: Run Pyrefly check
-        run: uv run pyrefly check scripts/*.py
-
-      - name: Run Python tests
-        run: uv run pytest scripts/test_generate_protocol.py
