+++ .vscode/settings.json
@@ -2,7 +2,6 @@
     "files.exclude": {
         "**/__pycache__": true,
         ".git": true,
-        ".pytest_cache": true,
         ".ruff_cache": true,
         ".venv": true,
