--- scripts/test_concept_errors.py
@@ -1,310 +1,0 @@
-"""
-Tests for compiler error messages when protocol concepts are violated.
-
-These tests verify that the C++ concepts and static asserts provide meaningful feedback.
-
-These tests are run from CMake using CTest and require flags to be passed at run-time.
-"""
-
-import os
-import re
-import subprocess
-import tempfile
-from typing import Any
-from typing import Callable
-from typing import List
-
-import pytest
-
-
-def pytest_addoption(parser: Any) -> None:
-    """Add command-line options for the compiler and flags."""
-    parser.addoption("--compiler", action="store", default="g++")
-    parser.addoption("--flags", action="append", default=[])
-
-
-@pytest.fixture
-def compiler(request: Any) -> str:
-    """Fixture that provides the compiler path."""
-    return str(request.config.getoption("--compiler"))
-
-
-@pytest.fixture
-def flags(request: Any) -> List[str]:
-    """Fixture that provides the compiler flags."""
-    return list(request.config.getoption("--flags"))
-
-
-def run_compiler(
-    compiler: str, flags: List[str], source_code: str
-) -> subprocess.CompletedProcess[str]:
-    """Run the compiler on the given source code."""
-    with tempfile.TemporaryDirectory() as tmpdir:
-        source_file = os.path.join(tmpdir, "test.cc")
-        obj_file = os.path.join(tmpdir, "test.o")
-
-        with open(source_file, "w") as f:
-            f.write(source_code)
-
-        cmd = [compiler] + flags + ["-c", source_file, "-o", obj_file]
-        return subprocess.run(cmd, capture_output=True, text=True)
-
-
-@pytest.fixture
-def compile_check(compiler: str, flags: List[str]) -> Callable[[str, List[str]], None]:
-    """Fixture that provides a function to check compiler errors."""
-
-    def check(source: str, expected_patterns: List[str]) -> None:
-        res = run_compiler(compiler, flags, source)
-        assert res.returncode != 0, "Compilation should have failed"
-        output = res.stderr + res.stdout
-        for pattern in expected_patterns:
-            assert re.search(pattern, output), (
-                f"Expected pattern '{pattern}' not found in compiler output.\n"
-                "--- Compiler Output ---\n"
-                f"{output}\n"
-                "-----------------------"
-            )
-
-    return check
-
-
-def test_missing_method(compile_check: Callable[[str, List[str]], None]) -> None:
-    """
-    Test that a class missing a required method fails.
-
-    The class fails to satisfy the protocol concept.
-    """
-    source = """
-    #include "generated/protocol_A.h"
-    #include "interface_A.h"
-    #include <utility>
-
-    class BadALike_MissingMethod {
-    public:
-        int count() { return 42; }
-    };
-
-    void test() {
-        xyz::protocol<xyz::A> a(std::in_place_type<BadALike_MissingMethod>);
-    }
-    """
-    compile_check(
-        source,
-        [
-            r"protocol_concept_A",
-            r".name\(\)",
-        ],
-    )
-
-
-def test_wrong_return_type(compile_check: Callable[[str, List[str]], None]) -> None:
-    """
-    Test that a class with a method having a wrong return type fails.
-
-    The class fails to satisfy the protocol concept.
-    """
-    source = """
-    #include "generated/protocol_A.h"
-    #include "interface_A.h"
-    #include <utility>
-    #include <string>
-
-    class BadALike_WrongReturnType {
-    public:
-        std::string_view name() const noexcept { return "name"; }
-        std::string count() { return "42"; }  // not the same type as int
-    };
-
-    void test() {
-        xyz::protocol<xyz::A> a(std::in_place_type<BadALike_WrongReturnType>);
-    }
-    """
-    compile_check(source, [r"protocol_concept_A", r"count\(\)"])
-
-
-def test_primary_template_instantiation(
-    compile_check: Callable[[str, List[str]], None],
-) -> None:
-    """Test that the primary protocol template cannot be instantiated."""
-    source = """
-    #include "protocol.h"
-    struct NoSpecialization {};
-    void test() { xyz::protocol<NoSpecialization> p; }
-    """
-    compile_check(
-        source,
-        [
-            r"static assertion failed",
-            r"The primary xyz::protocol template cannot be instantiated",
-        ],
-    )
-
-
-def test_view_const_to_mutable_concrete(
-    compile_check: Callable[[str, List[str]], None],
-) -> None:
-    """
-    Test that a protocol_view for a mutable interface cannot be constructed.
-
-    Fails when constructed from a const object.
-    """
-    source = """
-    #include "generated/protocol_A.h"
-    #include "interface_A.h"
-
-    struct MutALike {
-        std::string_view name() const noexcept { return "name"; }
-        int count() { return 1; } // Non-const
-    };
-
-    void test() {
-        const MutALike a;
-        xyz::protocol_view<xyz::A> view(a);
-    }
-    """
-    # The error should indicate that protocol_concept_A is not satisfied
-    # because count() is not const
-    compile_check(source, [r"protocol_concept_A", r"t.count\(\)"])
-
-
-def test_view_const_to_mutable_protocol(
-    compile_check: Callable[[str, List[str]], None],
-) -> None:
-    """
-    Test that a protocol_view for a mutable interface cannot be constructed.
-
-    Fails when constructed from a const protocol.
-    """
-    source = """
-    #include "generated/protocol_A.h"
-    #include "interface_A.h"
-
-    void test() {
-        const xyz::protocol<xyz::A> a;
-        xyz::protocol_view<xyz::A> view(a);
-    }
-    """
-    compile_check(source, [r"protocol_concept_A", r"t.count\(\)"])
-
-
-def test_view_const_alike_to_mutable(
-    compile_check: Callable[[str, List[str]], None],
-) -> None:
-    """
-    Test that a protocol_view for a mutable interface cannot be constructed.
-
-    Fails when constructed from an object missing mutable methods.
-    """
-    source = """
-    #include "generated/protocol_A.h"
-    #include "interface_A.h"
-
-    struct ConstALike {
-        std::string_view name() const noexcept { return "name"; }
-    };
-
-    void test() {
-        ConstALike a;
-        xyz::protocol_view<xyz::A> view(a);
-    }
-    """
-    compile_check(source, [r"protocol_concept_A", r"t.count\(\)"])
-
-
-def test_noexcept_violation(compile_check: Callable[[str, List[str]], None]) -> None:
-    """
-    Test that a class with a method missing noexcept fails.
-
-    The class fails to satisfy the protocol concept.
-    """
-    source = """
-    #include "generated/protocol_A.h"
-    #include "interface_A.h"
-    #include <utility>
-
-    class BadALike_NoExceptViolation {
-    public:
-        std::string_view name() const { return "name"; } // Missing noexcept
-        int count() { return 42; }
-    };
-
-    void test() {
-        xyz::protocol<xyz::A> a(std::in_place_type<BadALike_NoExceptViolation>);
-    }
-    """
-    compile_check(source, [r"protocol_concept_A", r"name\(\)"])
-
-
-def test_invalid_view_widening(compile_check: Callable[[str, List[str]], None]) -> None:
-    """
-    Verify that a protocol_view cannot be widened to a protocol_view with more methods.
-
-    Specifically, from A_Subset to A.
-    """
-    source = """
-    #include "generated/protocol_A.h"
-    #include "generated/protocol_A_Subset.h"
-    #include "interface_A.h"
-    #include "interface_A_Subset.h"
-
-    void test(xyz::protocol_view<xyz::A_Subset> view_subset) {
-        xyz::protocol_view<xyz::A> view_a(view_subset);
-    }
-    """
-    compile_check(source, [r"no member named|no matching function for call"])
-
-
-def test_invalid_view_constness_narrowing(
-    compile_check: Callable[[str, List[str]], None],
-) -> None:
-    """
-    Verify that a const protocol_view cannot be converted to a mutable protocol_view.
-
-    Even with narrowing (e.g. from const A to A_Subset).
-    """
-    source = """
-    #include "generated/protocol_A.h"
-    #include "generated/protocol_A_Subset.h"
-    #include "interface_A.h"
-    #include "interface_A_Subset.h"
-
-    void test(xyz::protocol_view<const xyz::A> view_const_a) {
-        xyz::protocol_view<xyz::A_Subset> view_subset(view_const_a);
-    }
-    """
-    compile_check(
-        source,
-        [
-            r"no matching function for call|no matching constructor|"
-            r"cannot convert|no matching template for|template argument deduction"
-        ],
-    )
-
-
-def test_invalid_protocol_widening(
-    compile_check: Callable[[str, List[str]], None],
-) -> None:
-    """
-    Verify that an owning protocol cannot be widened to a protocol with more methods.
-
-    Specifically, from A_Subset to A.
-    """
-    source = """
-    #include "generated/protocol_A.h"
-    #include "generated/protocol_A_Subset.h"
-    #include "interface_A.h"
-    #include "interface_A_Subset.h"
-
-    void test(xyz::protocol<xyz::A_Subset, std::allocator<std::byte>> p_subset) {
-        xyz::protocol<xyz::A, std::allocator<std::byte>> p(std::move(p_subset));
-    }
-    """
-    compile_check(
-        source,
-        [
-            r"no matching function for call|no matching constructor|"
-            r"cannot convert|static assertion failed|no matching template for|"
-            r"template argument deduction|no member named|has no member"
-        ],
-    )
